The laser cut version of the chassis
This chassis is meant for laser cut fabrication in wood or plastic, for default plate thickness of 5mm. However this can be set in the scad file. A few parts are meant to be 3D printed. In future versions, i'll try to make alternative parts so that it can be fabricated without a 3D printer.
By printing a paper version, it should be possible to attach it to a piece of wood and cut the shapes, and drill the holes with basic tools. i'll try to keep this version easy to make by hand, or make an alternative easier version.