Newer 3D printing only version of the chassis
This scad files contains what is required to have a 4 Lipo cells
batteries chassis, with 30T pololu tracks and motors. The chassis is
made of one part, containing all of this. On top of it, an electronics
fixation plate gets fixed by screws (it is also fully 3D printed). All
the files are meant to be easily printed, without support.
More about this here :
https://hackaday.io/project/3959-rhasika-a-simple-yet-powerful-robot-base-rpi/log/32224-new-chassis-design-and-new-batteries